break up typeDefs into separate modules
